as-connection-service/requirements.txt
fastapi==0.104.1
python-socketio==5.11.0
aioredis==2.0.1
python-jose[cryptography]==3.3.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Shared dependencies
../shared